chap4

+++ chap-04/c.php
@@ -3,7 +3,7 @@
     $score = $_POST["score"];
     $result;
 
-    switch ($score / 10) {
+    switch ((int)($score / 10)) {
       case 10: 
       case 9: $result = "A"; break;
       case 8: $result = "B"; break;
